Moved validation functionality to separate file in lib/ folder (to prevent bug with deleted field validation in another PR). Added validation to 'description' field
